Modularize cloud ASR providers
