+++ .env.example
@@ -211,6 +211,15 @@ ASR_GLADIA_FILE_LIMIT_MIN_CHUNK_SEC=600
 ASR_GLADIA_FILE_LIMIT_MAX_CHUNKS=200
 ASR_GLADIA_AUDIO_BITRATE=32k
 
+# Gemini cloud ASR local free-tier guard. Keep enabled to smooth requests before
+# the provider rejects them; set ASR_GEMINI_FREE_TIER_LIMITER_ENABLED=0 to rely
+# on Gemini's live quota errors only. Override RPM/TPM/RPD in local .env to
+# match the selected Gemini model and tier.
+ASR_GEMINI_FREE_TIER_LIMITER_ENABLED=1
+ASR_GEMINI_FREE_TIER_RPM=5
+ASR_GEMINI_FREE_TIER_TPM=250000
+ASR_GEMINI_FREE_TIER_RPD=20
+
 # ============================================================================
 # 6) OpenVINO Runtime Controls (common paths, helpers, and timeouts)
 # ============================================================================
